Reparando los repository

+++ src/main/java/com/grupo14/turnos/repository/EmpleadoRepository.java
@@ -3,5 +3,5 @@ package com.grupo14.turnos.repository;
 import com.grupo14.turnos.modelo.Empleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmpleadoRepository extends JpaRepository<Empleado, Integer> {
+public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
 }
