+++ src/main/java/com/grupo14/turnos/repository/UsuarioRepository.java
@@ -1,11 +1,10 @@
 package com.grupo14.turnos.repository;
 
 import com.grupo14.turnos.modelo.Usuario;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
     Optional<Usuario> findByEmail(String email);
 }
